Correct the desynchronization bug in the pagination between delete and version markers
Pages returned by the boto3 paginator contain both object version
markers and delete markers. It can happen that a version marker and its
corresponding delete markers are on two consecutive pages, and not in
the same one as the code expects.

We have only seen this happen for version markers which are at
the top of the their page (element 0 of the version list), and in each
case, the delete marker was on the immediately anterior page.

The observable consequence is that a restore operation can sometimes
restore a file that should not be restored. When this happens, several
consecutive identical runs of the restore commands will restore
different additional files, depending on how the frontiers of the page
are decided by boto3.

In order to solve the issue, the proposed patch "pushes" the delete
markers that were not relevant for the version markers of the page, to
the next page where they may be needed.